add schedule movie impl
