+++ .idea/gradle.xml
@@ -17,8 +17,8 @@
             <option value="$PROJECT_DIR$/feature_detail" />
             <option value="$PROJECT_DIR$/feature_main" />
             <option value="$PROJECT_DIR$/feature_search" />
+            <option value="$PROJECT_DIR$/tasks" />
             <option value="$PROJECT_DIR$/utils" />
-            <option value="$PROJECT_DIR$/workers" />
           </set>
         </option>
         <option name="resolveModulePerSourceSet" value="false" />
