+++ .idea/misc.xml
@@ -55,6 +55,7 @@
         <entry key="..\:/Project/MoviesDemo/feature_search/src/main/res/drawable/ic_search.xml" value="0.28307692307692306" />
         <entry key="..\:/Project/MoviesDemo/feature_search/src/main/res/layout/fragment_search.xml" value="0.1" />
         <entry key="..\:/Project/MoviesDemo/feature_search/src/main/res/layout/movie_item.xml" value="0.2713541666666667" />
+        <entry key="..\:/Sdk/platform-tools/caches/transforms-3/0b86320c71311844fa2230d5c8600c8e/transformed/appcompat-1.3.1/res/drawable/abc_dialog_material_background.xml" value="0.2755208333333333" />
       </map>
     </option>
   </component>
