+++ README.md
@@ -3,4 +3,5 @@
 ## Preview
 <p align="left">
 <img src="img/movieDemo2.gif" width="32%"/>
+<img src="img/movieDemoSchedule.gif" width="32%"/>
 </p>
