+++ feature_detail/src/main/java/com/bosha/feature_detail/ui/DetailFragment.kt
@@ -17,16 +17,15 @@ import androidx.lifecycle.lifecycleScope
 import coil.imageLoader
 import coil.request.ImageRequest
 import com.bosha.domain.entities.MovieDetails
-import com.bosha.domain.utils.TaskScheduler
 import com.bosha.feature_detail.R
 import com.bosha.feature_detail.databinding.FragmentDetailBinding
+import com.bosha.feature_detail.utils.datetime.schedule
 import com.bosha.utils.extensions.doOnEndTransition
 import com.bosha.utils.navigation.Screens
 import com.google.android.material.transition.MaterialContainerTransform
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
-import java.time.Duration
 import javax.inject.Inject
 
 @AndroidEntryPoint
@@ -41,10 +40,6 @@ class DetailFragment : Fragment() {
 
     @Inject
     lateinit var viewModelAssistedFactory: DetailViewModel.Factory
-
-    //delete
-    @Inject
-    lateinit var scheduler: TaskScheduler
 
     private val viewModel by viewModels<DetailViewModel> {
         DetailViewModel.provideFactory(viewModelAssistedFactory, movieId)
@@ -115,12 +110,12 @@ class DetailFragment : Fragment() {
         tvRating.text = details.votes.toString()
 
         tvStory.text = details.overview
-        ////test
-        tvStory.setOnClickListener {
-            scheduler.scheduleNotification(details.id.toString(), Duration.ofSeconds(10))
-        }
-        ////
 
+        ibSchedule.setOnClickListener {
+            schedule { duration ->
+                viewModel.scheduleMovie(details.id.toString(), duration)
+            }
+        }
 
         tvRunningTime.text = "${details.runtime} min"
 
@@ -188,7 +183,9 @@ class DetailFragment : Fragment() {
             tvRunningTime,
             ibFavorite,
             gradient,
-            tvGenres
+            tvGenres,
+            ibSchedule,
+            tvSchedule
         )
     }
 }
