+++ feature_detail/src/main/res/values/strings.xml
@@ -3,4 +3,5 @@
     <string name="cast">Cast</string>
     <string name="storyline">Story line</string>
     <string name="transition_name">transition_name</string>
+    <string name="schedule_movie">Schedule movie</string>
 </resources>
