+++ settings.gradle
@@ -14,4 +14,4 @@ include ':domain'
 include ':data'
 include ':utils'
 include ':feature_detail'
-include ':workers'
+include ':tasks'
