--- workers/build.gradle
@@ -1,62 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-    id 'kotlin-kapt'
-    id 'dagger.hilt.android.plugin'
-}
-
-android {
-    compileSdk 31
-
-    defaultConfig {
-        minSdk 24
-        targetSdk 31
-        versionCode 1
-        versionName "1.0"
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles "consumer-rules.pro"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = '1.8'
-    }
-    buildFeatures {
-        viewBinding true
-    }
-}
-dependencies {
-    implementation project(':domain')
-    implementation project(':utils')
-    implementation project(':data')
-}
-dependencies {
-    implementation Libs.kotlinStdLib
-    implementation Libs.corektx
-    implementation Libs.appcompat
-    implementation Libs.material
-    implementation Libs.constraint
-    //coroutines
-    implementation Libs.coroutines
-    //hilt
-    implementation Libs.hilt
-    kapt Libs.hiltCompiler
-    //work manager
-    api Libs.workManager
-    api Libs.hiltWork
-    kapt Libs.androidxHiltCompiler
-    //coil
-    implementation Libs.coil
-
-    testImplementation Libs.junit
-}
